Begin repurposing a Collection for "My List"
A photo now has a donor, rather than belonging to a collection. Photos
will be able to belong to appear in 0 or more collections. The
collections will be Public, Private, or Unlisted (meaning they won't be
listed in the directory but you can see them if you have a link).