Fix for timeline not appearing
It's very strange, but the database query which annotates photos with
row numbers, so that a year can be matched to a given page number, was
returning the photos in no particular order. It then iterates from the
year of the first photo to the year of the last photo, which will
produce no iterations in the loop if the last photo is from an earlier
year than the first photo.