Support new simpler search across multiple fields
This allows users to simply type "house" and get photos that are either
tagged or termed with house. There are a few things that still need to
be done, but this has finally reached acceptable parity with the old
collection system. It still needs:

1. support for removing individual clauses from the filter.

2. Donors are not included in this search.

3. Autocomplete does not work.

4. Placing quotes around multiple words will not work as expected. This
is awkward because if it did, "Cedar Falls" would return photos from
Cedar Falls, but "Waterloo, IA" would not. It will need to either
recognize this format or have something sort of like a hidden tag to
make this work. Not sure.