HTMX extension for SVG tags
HTMX prevents default processing on things like clicks on links. The
default is to load the href. HTML is case insensitive, so Javascript
will see the event's element as having "A" as the tag name. SVG is case
sensitive, so Javascript sees "A" as the tag name. HTMX looks for "A"
and therefore does not prevent default processing on SVG links. This
extension prevents default processing on "a" click events.