Drag and drop of year marker works
It does not work on embedded site. This is because the drag-drop.js file
does not load due to being a link tag, and loading by other means would
most likely be a CORS violation. Therefore it should be turned into an
ES6 module and rolled into compiled and minified output. It should work
then.