Gridview does not propagate a search to timeline
This is a little bit confusing, but caption search is relevance ordered.
If it propagates its filter to the timeline view, photos that do not
match the search particularly well will be in the timeline. I think the
site could do a better job conveying how this works to users, as they
may not realize the results are not in chronological order and they may
not understand that they are going to lose their filtering.