Random expression generator
While neat, I don't think it has any use because the odds that it will
match any random photo are very low. It may be better to be able to ask
a photo what expressions it matches. Then could generate random
complex expressions A and B that match it. Then verify that the photo
matches both, that it does not match the negation of the expression,
that it matches A AND B, A OR B, that it matches (A AND B) or C where C
could be YearEquals(someOtherYear), that it does not match A AND B AND
C, and that it matches (A OR C) AND (B OR C).