Remove deal side effects/exception contracts
Deal.pure is deal.has() and deal.safe. The side effect markers are only
checked for a few things, and others are checked by the linter. This is
relatively harmless. On the other hand, deal.safe causes the exceptions
caused by bugs during debugging to produce useless and confusing error
messages. Therefore, I think deal.safe and deal.pure should not be used.
They're meant as documentation and to work with deal's tester.