351 per archive logos and other theme options
Also including minimal support for 367, which allows swapping the base
template on other archives. The alternate base templates could extend
the main base template and override specific blocks or use a totally
different layout. There are a few IDs that must exist for various
content to load. This turned out to not be useful for the theming, but
it was relatively simple to enable.