382 Combined tags/terms can be built from template
Photo objects now have a get_all_tags method that returns approved tags,
tags submitted by the current user, and terms. They are annotated with
label and label_lower to provide a consistent interface for sorting
(with dictsort) and generating html. There is also a template tag which
can be used to pass the user to the method.