Thumbnail carousel always has 20 thumbs on both sides
There will be spacers with class="empty" on the img tag if there are not
enough objects in the carousel. References to fortepan.us.position http
header have been removed. The carousel's next and previous links have
been removed from the context because they depend on the width of the
carousel and that is not known.