Fix hx-trigger middleware problem
This will work for now but HttpResponse objects should have a hxtrigger
property that is Union[None, str, Dict[str, Any]] type, and the str
value should be a singular event name (not json encoded events). The
header should not be set in view functions. The middleware can then
easily add more to it and encode it.